feat: pass lockfileId to pnpm-sync
